+++ NEWS.md
@@ -1,7 +1,7 @@
 # eatGADS 0.19.1.9000
 
 * `checkUniqueness2()` for faster checking if a variable is unique within an identifier varible
-* `fixEncoding()` for fixing encoding issues caused during the import via `import_spss()` using `ASCII` encoding
+* `fixEncoding()` for fixing encoding issues caused during the import via `import_spss()` using `ASCII` or other encoding
 * `composeVar()` for combining the information of two variables in a third variable
 * `subImputations()` now works with differing `varName` and `varName_imp`
 * `fillImputations()` for efficiently filling missing values in imputed variables via not imputed variables